pyproject.toml
[project]
name = "fichero-printer"
version = "0.1.0"
description = "Fichero D11s thermal label printer - BLE CLI tool"
requires-python = ">=3.10"
dependencies = [
    "bleak",
    "pillow",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["fichero"]

[project.scripts]
fichero = "fichero.cli:main"
